+++ playground/vite-project/postcss.config.cjs
@@ -1,0 +1,12 @@
+module.exports = {
+  plugins: {
+    'tailwindcss/nesting': {},
+    'tailwindcss': {},
+    '@minko-fe/postcss-pxtoviewport': {
+      replace: true,
+      minPixelValue: 1,
+      atRules: false,
+      propList: ['*'],
+    },
+  },
+}